third_party/WebKit/LayoutTests/fast/css/inherited-properties-explicit-expected.txt

PASS test('test1a') is "0px"
PASS test('test1b') is "0px"
PASS test('test2a') is "5px"
PASS test('test2b') is "5px"
PASS test('test3a') is "10px"
PASS test('test3b') is "10px"
PASS successfullyParsed is true

TEST COMPLETE


Test that the matched declaration cache handles explicitly inherited properties correctly.
